Creación de las clases de implementación de servicios: FuncionServiceImpl, PagoServiceImpl, PeliculaServiceImpl, SalaServiceImpl y VentaServiceImpl

+++ src/main/java/com/example/CineSpringBoot/repositories/SalaRepository.java
@@ -4,5 +4,5 @@ import com.example.CineSpringBoot.entities.Sala;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SalaRepository extends BaseRepository<Sala, Integer> {
+public interface SalaRepository extends BaseRepository<Sala, Long> {
 }
