+++ src/main/java/com/example/CineSpringBoot/services/PeliculaService.java
@@ -2,5 +2,5 @@ package com.example.CineSpringBoot.services;
 
 import com.example.CineSpringBoot.entities.Pelicula;
 
-public interface PeliculaService extends BaseService<Pelicula, Integer> {
+public interface PeliculaService extends BaseService<Pelicula, Long> {
 }
